Restructure extrac into wllvm and add docs
