+++ tests/test_settings.py
@@ -28,7 +28,9 @@ class BenchBuildConfig(unittest.TestCase):
     def tearDownClass(cls):
         cls.tmp_file.close()
 
-    def check_all_files_in_config_list(self, folder, config_list,
+    def check_all_files_in_config_list(self,
+                                       folder,
+                                       config_list,
                                        exclude_list=None):
         """
         Check if all python files in a folder are added to the
@@ -76,11 +78,9 @@ class BenchBuildConfig(unittest.TestCase):
 
         loaded_plugins = BB_CFG["plugins"]["projects"].value
         self.check_all_files_in_config_list("varats/projects/c_projects/",
-                                            loaded_plugins,
-                                            excluded_projects)
+                                            loaded_plugins, excluded_projects)
         self.check_all_files_in_config_list("varats/projects/cpp_projects/",
-                                            loaded_plugins,
-                                            excluded_projects)
+                                            loaded_plugins, excluded_projects)
 
     def test_if_experiments_where_added(self):
         """
@@ -88,7 +88,7 @@ class BenchBuildConfig(unittest.TestCase):
         """
         excluded_experiments = [
             "wllvm.py", "phasar.py", "region_instrumentation.py",
-            "commit_annotation_report.py", "blame_experiment.py", "extract.py"
+            "commit_annotation_report.py", "blame_experiment.py"
         ]
 
         loaded_plugins = BB_CFG["plugins"]["experiments"].value
