+++ varats/experiments/phasar_env_analysis.py
@@ -19,13 +19,13 @@ from benchbuild.project import Project
 from benchbuild.experiment import Experiment
 import benchbuild.utils.actions as actions
 from benchbuild.utils.cmd import phasar, mkdir, timeout
+
 from varats.data.reports.env_trace_report import EnvTraceReport as ENVR
 from varats.data.report import FileStatusExtension as FSE
-from varats.experiments.extract import Extract
-from varats.experiments.wllvm import RunWLLVM
-from varats.utils.experiment_util import (
-    exec_func_with_pe_error_handler, FunctionPEErrorWrapper,
-    PEErrorHandler)
+from varats.experiments.wllvm import RunWLLVM, Extract
+from varats.utils.experiment_util import (exec_func_with_pe_error_handler,
+                                          FunctionPEErrorWrapper,
+                                          PEErrorHandler)
 
 
 class PhasarEnvIFDS(actions.Step):  # type: ignore
@@ -41,8 +41,7 @@ class PhasarEnvIFDS(actions.Step):  # type: ignore
     RESULT_FOLDER_TEMPLATE = "{result_dir}/{project_dir}"
 
     def __init__(self, project: Project):
-        super(PhasarEnvIFDS, self).__init__(
-            obj=project, action_fn=self.analyze)
+        super(PhasarEnvIFDS, self).__init__(obj=project, action_fn=self.analyze)
 
     def analyze(self) -> actions.StepResult:
         """
@@ -80,12 +79,12 @@ class PhasarEnvIFDS(actions.Step):  # type: ignore
                 project_version=str(project.version))
 
             # Define result file.
-            result_file = ENVR.get_file_name(
-                project_name=str(project.name),
-                binary_name=binary.name,
-                project_version=str(project.version),
-                project_uuid=str(project.run_uuid),
-                extension_type=FSE.Success)
+            result_file = ENVR.get_file_name(project_name=str(project.name),
+                                             binary_name=binary.name,
+                                             project_version=str(
+                                                 project.version),
+                                             project_uuid=str(project.run_uuid),
+                                             extension_type=FSE.Success)
 
             # Define output file name of failed runs
             error_file = ENVR.get_file_name(project_name=str(project.name),
@@ -97,19 +96,18 @@ class PhasarEnvIFDS(actions.Step):  # type: ignore
                                             file_ext=".txt")
 
             # Put together the run command
-            phasar_run_cmd = phasar["-D", "IFDS_EnvironmentVariableTracing",
-                                    "-m", "{cache_folder}/{bc_file}"
-                                    .format(cache_folder=bc_cache_dir,
-                                            bc_file=bc_target_file),
-                                    "-O", "{res_folder}/{res_file}".format(
-                                        res_folder=result_folder,
-                                        res_file=result_file)]
+            phasar_run_cmd = phasar[
+                "-D", "IFDS_EnvironmentVariableTracing", "-m",
+                "{cache_folder}/{bc_file}".
+                format(cache_folder=bc_cache_dir, bc_file=bc_target_file), "-O",
+                "{res_folder}/{res_file}".
+                format(res_folder=result_folder, res_file=result_file)]
 
             # Run the phasar command with custom error handler and timeout
             exec_func_with_pe_error_handler(
                 timeout[timeout_duration, phasar_run_cmd],
-                PEErrorHandler(result_folder, error_file,
-                               phasar_run_cmd, timeout_duration))
+                PEErrorHandler(result_folder, error_file, phasar_run_cmd,
+                               timeout_duration))
 
 
 class UnlimitStackSize(actions.Step):  # type: ignore
@@ -122,8 +120,8 @@ class UnlimitStackSize(actions.Step):  # type: ignore
     DESCRIPTION = "Sets new resource limits."
 
     def __init__(self, project: Project):
-        super(UnlimitStackSize, self).__init__(
-            obj=project, action_fn=self.__call__)
+        super(UnlimitStackSize, self).__init__(obj=project,
+                                               action_fn=self.__call__)
 
     def __call__(self) -> actions.StepResult:
         """
@@ -164,12 +162,11 @@ class PhasarEnvironmentTracing(Experiment):  # type: ignore
                 PhasarEnvIFDS.RESULT_FOLDER_TEMPLATE.format(
                     result_dir=str(BB_CFG["varats"]["outfile"]),
                     project_dir=str(project.name)),
-                ENVR.get_file_name(
-                    project_name=str(project.name),
-                    binary_name="all",
-                    project_version=str(project.version),
-                    project_uuid=str(project.run_uuid),
-                    extension_type=FSE.CompileError)))
+                ENVR.get_file_name(project_name=str(project.name),
+                                   binary_name="all",
+                                   project_version=str(project.version),
+                                   project_uuid=str(project.run_uuid),
+                                   extension_type=FSE.CompileError)))
 
         analysis_actions = []
